+++ src/lib/utils/merge-content.ts
@@ -1,0 +1,7 @@
+import { escapeRegExp } from "./regexp";
+
+export function mergeContent(newEntryHref: string, newEntry: string, existinContent: string): string {
+  const existingItemPattern = String.raw`^- \[.+\]\(${escapeRegExp(newEntryHref)}\).*$`;
+  const replaceResult = existinContent.replace(new RegExp(existingItemPattern, "m"), newEntry);
+  return replaceResult === existinContent ? `${newEntry}\n${existinContent}` : replaceResult;
+}