+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # v3.0.0
 
 - New: Migrated to Web Manifest v3 for compatibility with Chrome
+- New: Detect existing URLs for which "Save" becomes "Update"
+- Changed: Link input is moved above Title input
 
 # v2.3.3